[BE] fix: Highlight 엔티티의 데이터 타입을 db 테이블와 일치시킨다. (#829)
* fix: db와 데이터 타입 범위 일치를 위해 필드 타입 int을 변경

* refactor: dto의 인덱스 관련 데이터 타입을 int로 변경

* refactor: null을 처리하도록 DTO의 필드 데이터 타입을 Integer로 변경

* refactor: 불필요한 타입 캐스팅 제거

+++ backend/src/main/java/reviewme/highlight/domain/Highlight.java
@@ -29,7 +29,7 @@ public class Highlight {
     @Embedded
     private HighlightPosition highlightPosition;
 
-    public Highlight(long answerId, long lineIndex, long startIndex, long endIndex) {
+    public Highlight(long answerId, int lineIndex, int startIndex, int endIndex) {
         this.answerId = answerId;
         this.highlightPosition = new HighlightPosition(lineIndex, startIndex, endIndex);
     }
