+++ backend/src/main/java/reviewme/highlight/service/dto/HighlightedLineRequest.java
@@ -8,7 +8,7 @@ import java.util.List;
 public record HighlightedLineRequest(
 
         @NotNull(message = "인덱스를 입력해주세요.")
-        Long index,
+        Integer index,
 
         @Valid @NotEmpty(message = "하이라이트 범위를 입력해주세요.")
         List<HighlightIndexRangeRequest> ranges
